requirements.txt
streamlit==1.40.0
google-generativeai>=0.8.0

